bring tests back to the typescript
