+++ chrome-launcher/test/chrome-launcher-test.ts
@@ -5,18 +5,32 @@
  */
 'use strict';
 
-import {Launcher} from '../chrome-launcher';
+import {Launcher, ProtocolPortDetails} from '../chrome-launcher';
 import {spy, stub} from 'sinon';
 import * as assert from 'assert';
 import * as stream from 'stream';
+import * as fs from 'fs';
 
 const log = require('lighthouse-logger');
-const fsMock = {
+
+const fsMock = Object.assign({}, fs, {
   writeFileSync: () => {},
   createWriteStream: () => {
     return new stream.Writable();
   }
-};
+});
+
+function createMockChildProcess() {
+  // `as any` used to avoid TS private member errors
+  const mockedStream = new stream.Readable() as any;
+  mockedStream._read = () => {};
+
+  return {
+    pid: 'some_pid',
+    stdout: mockedStream,
+    stderr: mockedStream
+  };
+}
 
 describe('Launcher', () => {
   beforeEach(() => {
@@ -26,18 +40,6 @@ describe('Launcher', () => {
   afterEach(() => {
     log.setLevel('');
   });
-
-  function createMockChildProcess() {
-    const mockedStream = new stream.Readable();
-    Object.defineProperty(mockedStream, '_read', {
-      value: (size:number) => { size; }
-    });
-    return {
-      pid: 'some_pid',
-      stdout: mockedStream,
-      stderr: mockedStream
-    };
-  }
 
   it('sets default launching flags', async () => {
 
@@ -104,7 +106,6 @@ describe('Launcher', () => {
 
   it('doesn\'t fail when killed twice', async () => {
     const chromeInstance = new Launcher();
-
     await chromeInstance.launch();
     await chromeInstance.kill();
     await chromeInstance.kill();
@@ -140,4 +141,81 @@ describe('Launcher', () => {
     const chromeFlags = spawnStub.getCall(0).args[1] as string[];
     assert.ok(!chromeFlags.includes('--disable-extensions'));
   }).timeout(2000);
+});
+
+describe('Launcher features', () => {
+
+
+  describe('getActivePort', () => {
+
+    it('can handle Chrome >=62', async () => {
+      const mockChrome = createMockChildProcess();
+      // `as any` used to avoid TS private member errors
+      const LauncherClass = Launcher as any;
+      const data = await LauncherClass.prototype.getActivePort(mockChrome) as ProtocolPortDetails;
+
+      assert.deepStrictEqual(data, {
+        port: 63567,
+        browserWs: 'ws://127.0.0.1:63567/devtools/browser/2f168fe0-2d64-48aa-a22c-6ccaff6e9b24'
+      });
+
+      mockChrome.stderr.emit('data', 'DevTools listening on ws://127.0.0.1:63567/devtools/browser/2f168fe0-2d64-48aa-a22c-6ccaff6e9b24');
+      mockChrome.stderr.emit('end');
+    });
+
+    it('can handle Chrome <=61', async () => {
+      const mockChrome = createMockChildProcess();
+      const LauncherClass = Launcher as any;
+      const data = await LauncherClass.prototype.getActivePort(mockChrome) as ProtocolPortDetails;
+      assert.deepStrictEqual(data, {
+        port: 64223,
+        browserWs: null
+      });
+
+      mockChrome.stderr.emit('data', 'DevTools listening on 127.0.0.1:64223');
+      mockChrome.stderr.emit('end');
+    });
+
+    it('can handle Chrome dev build (<= 61)', async () => {
+      const mockChrome = createMockChildProcess();
+      const LauncherClass = Launcher as any;
+      const data = await LauncherClass.prototype.getActivePort(mockChrome) as ProtocolPortDetails;
+      assert.deepStrictEqual(data, {
+        port: 57747,
+        browserWs: null
+      });
+
+      mockChrome.stderr.emit(
+        'data',
+        `[50786:50695:0808/133113.866481:ERROR:devtools_http_handler.cc(786)]
+            DevTools listening on 127.0.0.1:57747
+
+            `
+      );
+      mockChrome.stderr.emit('end');
+    });
+  });
+
+  describe('isDebuggerReady', () => {
+    it('tries to first connect to a specific port if provided', async () => {
+      const requestedPort = 6789;
+
+      const isReadySpy = spy();
+      const prepareSpy = spy();
+      const spawnStub = stub().returns(createMockChildProcess());
+
+      const chromeInstance = new Launcher(
+          {port: requestedPort, enableExtensions: true},
+          {fs: fsMock, rimraf: spy(), spawn: spawnStub}) as any;
+
+      chromeInstance.isDebuggerReady = isReadySpy.bind(chromeInstance);
+      chromeInstance.prepare = prepareSpy.bind(chromeInstance);
+      stub(chromeInstance, 'waitUntilReady').returns(Promise.resolve());
+      stub(chromeInstance, 'getActivePort').returns(Promise.resolve());
+
+      await chromeInstance.launch();
+      assert.ok(isReadySpy.alwaysCalledWithExactly(6789), isReadySpy.args.toString());
+      assert.ok(isReadySpy.calledBefore(prepareSpy), 'isDebuggerReady not called before prepare');
+    }).timeout(2000);
+  });
 });
